+++ CHANGELOG.md
@@ -5,6 +5,7 @@ All notable changes to this project will be documented in this file.
 ## Unreleased
 
 - added event methods to highlight absolute block positions and rotation-aware offsets
+- added event to modify the conditions tooltip in recipe viewers
 
 ## [3.12.1] - 2026-04-26